requirements.txt
google-genai
python-dotenv
scikit-learn
faiss-cpu